uploaded text feature extractor script
This is a simple script that can be used to extract text features for use in other models. The script outputs a csv file that is sorted by the column 'difference' which is each row in the TP (True Positive) column minus the TN (True Negative) column. The tokens on the top of the list are most associated with TP while the bottom is most associated with TN.